Refactor: Cleanup and restructure
