+++ src/app/(auth)/components/AuthFormWrapper.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/utils/classnames';
 
 interface AuthFormWrapperProps {
 	children: ReactNode;
