+++ src/components/ui/label.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 
 import * as LabelPrimitive from '@radix-ui/react-label';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/utils/classnames';
 
 function Label({ className, ...props }: React.ComponentProps<typeof LabelPrimitive.Root>) {
 	return (
