+++ src/hooks/data/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Data hooks for server state management using TanStack Query.
+ * These hooks provide typed, cached access to API endpoints with loading states.
+ */
+
+export { useSessionUser } from './useSessionUser';
