+++ src/hooks/index.ts
@@ -1,1 +1,1 @@
-// Placeholder for hooks barrel file
+export * from './data';
