+++ src/lib/models/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Domain models and interfaces for the LangRoute application.
+ * Contains TypeScript types that represent core business entities
+ * and are shared across the frontend and backend layers.
+ */
+
+// User-related domain models
+export * from './User';
+
+// Authentication-related domain models
+export * from './Auth';
