+++ src/lib/types/global.d.ts
@@ -1,0 +1,1 @@
+//Placeholder for global TypeScript types
