+++ src/services/index.ts
@@ -1,0 +1,4 @@
+export { AuthService } from './auth/service';
+export * from './auth/service';
+export * as SystemError from './system/errorService';
+export { createErrorResponse, handleApiError } from './system/errorService'; // Export error handling functions
